hosted/libusb: Move platform.c libusb related functions to own file.
