+++ src/platforms/hosted/ftdi_bmp.c
@@ -28,6 +28,7 @@
 #include <sys/time.h>
 
 #include "ftdi_bmp.h"
+#include <libftdi1/ftdi.h>
 
 struct ftdi_context *ftdic;
